Implement review feedback KAN-96

+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/repository/IngredientRepository.kt
@@ -9,6 +9,5 @@ import java.util.*
 interface IngredientRepository : JpaRepository<Ingredient, Long> {
     fun findAllByEventId(eventId: Long): List<Ingredient>
     fun findByIdAndEventId(id: Long, eventId: Long): Optional<Ingredient>
-
-    fun findByIdIn(id: List<Int>): List<Ingredient>
+    fun findByIdIn(id: List<Long>): List<Ingredient>
 }
